# contract_snapshot/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "ContractCompatibilityRequest": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version"
        ],
        "title": "ContractCompatibilityRequest",
        "type": "object"
      },
      "ContractCompatibilityResult": {
        "additionalProperties": false,
        "properties": {
          "current_version": {
            "title": "Current Version",
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "migration_required": {
            "default": false,
            "title": "Migration Required",
            "type": "boolean"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "accepted_with_warnings",
              "unsupported_version"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "status",
          "current_version"
        ],
        "title": "ContractCompatibilityResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalDataBundleValidationRequest": {
        "additionalProperties": false,
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/LocalDataBundle"
          }
        },
        "required": [
          "bundle"
        ],
        "title": "LocalDataBundleValidationRequest",
        "type": "object"
      },
      "LocalDataBundleValidationResult": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "observation_results": {
            "items": {
              "$ref": "#/components/schemas/ObservationValidationResult"
            },
            "title": "Observation Results",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "status": {
            "default": "accepted",
            "enum": [
              "accepted",
              "accepted_with_warnings",
              "unsupported_version"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "ok"
        ],
        "title": "LocalDataBundleValidationResult",
        "type": "object"
      },
...
      "LocalStoreSnapshotValidationRequest": {
        "additionalProperties": false,
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/LocalStoreSnapshot"
          }
        },
        "required": [
          "snapshot"
        ],
        "title": "LocalStoreSnapshotValidationRequest",
        "type": "object"
      },
...
      "ObservationValidationRequest": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "$ref": "#/components/schemas/ObservationEvent"
          }
        },
        "required": [
          "event"
        ],
        "title": "ObservationValidationRequest",
        "type": "object"
      },
      "ObservationValidationResult": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "tracker_code": {
            "title": "Tracker Code",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "tracker_code"
        ],
        "title": "ObservationValidationResult",
        "type": "object"
      },
...
      "PcosEvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "evaluated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional anchor time for the rolling evaluation window. Defaults to now.",
            "title": "Evaluated At"
          },
          "evaluation_window_days": {
            "default": 365,
            "description": "Trailing window (days) of observations used for cycle and feature derivation.",
            "maximum": 1825.0,
            "minimum": 60.0,
            "title": "Evaluation Window Days",
            "type": "integer"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationEvent"
            },
            "title": "Observations",
            "type": "array"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "years_since_menarche": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional context used to apply 2023 International Evidence-based Guideline cycle-irregularity rules by reproductive age. Adolescent rules apply when <3 years post-menarche.",
            "title": "Years Since Menarche"
          }
        },
        "required": [
          "subject_id"
        ],
        "title": "PcosEvaluationRequest",
        "type": "object"
      },
      "PcosEvaluationResult": {
        "additionalProperties": false,
        "properties": {
          "adolescent_both_features_required": {
            "default": false,
            "title": "Adolescent Both Features Required",
            "type": "boolean"
          },
          "age_group_used": {
            "enum": [
              "adolescent_pubertal",
              "adolescent_post",
              "adult",
              "unknown"
            ],
            "title": "Age Group Used",
            "type": "string"
          },
          "analyzer_code": {
            "const": "pcos_feature_pattern_v1",
            "title": "Analyzer Code",
            "type": "string"
          },
          "analyzer_version": {
            "title": "Analyzer Version",
            "type": "string"
          },
          "confidence": {
            "enum": [
              "none",
              "low",
              "moderate",
              "high"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "cycle_irregularity": {
            "$ref": "#/components/schemas/PcosCycleIrregularitySignal"
          },
          "diagnostic_framework": {
            "const": "rotterdam_2003_via_2023_ieg",
            "default": "rotterdam_2003_via_2023_ieg",
            "title": "Diagnostic Framework",
            "type": "string"
          },
          "differential_reminders": {
            "items": {
              "type": "string"
            },
            "title": "Differential Reminders",
            "type": "array"
          },
          "evaluation_window_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Window End"
          },
          "evaluation_window_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Window Start"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_summary": {
            "title": "Evidence Summary",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "hyperandrogenism": {
            "$ref": "#/components/schemas/PcosHyperandrogenismSignal"
          },
          "meets_hyperandrogenism_rule": {
            "default": false,
            "title": "Meets Hyperandrogenism Rule",
            "type": "boolean"
          },
          "meets_irregularity_rule": {
            "default": false,
            "title": "Meets Irregularity Rule",
            "type": "boolean"
          },
          "metabolic_context": {
            "$ref": "#/components/schemas/PcosMetabolicSignal"
          },
          "pcom_assessment_could_resolve": {
            "default": false,
            "title": "Pcom Assessment Could Resolve",
            "type": "boolean"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "rotterdam_self_report_feature_count": {
            "default": 0,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Rotterdam Self Report Feature Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "insufficient_data",
              "suppressed",
              "features_absent",
              "features_partial",
              "features_present"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "suppressors": {
            "items": {
              "type": "string"
            },
            "title": "Suppressors",
            "type": "array"
          }
        },
        "required": [
          "analyzer_code",
          "analyzer_version",
          "subject_id",
          "generated_at",
          "status",
          "confidence",
          "age_group_used",
          "cycle_irregularity",
          "hyperandrogenism",
          "metabolic_context",
          "summary",
          "evidence_summary"
        ],
        "title": "PcosEvaluationResult",
        "type": "object"
      },
...
      "PmddEvaluationRequest": {
        "additionalProperties": false,
        "properties": {
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationEvent"
            },
            "title": "Observations",
            "type": "array"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "subject_id"
        ],
        "title": "PmddEvaluationRequest",
        "type": "object"
      },
      "PmddEvaluationResult": {
        "additionalProperties": false,
        "properties": {
          "activation_score": {
            "default": 0,
            "minimum": 0.0,
            "title": "Activation Score",
            "type": "number"
          },
          "analyzer_code": {
            "const": "pmdd_pattern_v1",
            "title": "Analyzer Code",
            "type": "string"
          },
          "analyzer_version": {
            "title": "Analyzer Version",
            "type": "string"
          },
          "confidence": {
            "enum": [
              "none",
              "low",
              "moderate",
              "high"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "coverage": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Coverage",
            "type": "number"
          },
          "cycle_evaluations": {
            "items": {
              "$ref": "#/components/schemas/PmddCycleEvaluation"
            },
            "title": "Cycle Evaluations",
            "type": "array"
          },
          "diagnostic_framework": {
            "const": "drsp_cpass_like",
            "default": "drsp_cpass_like",
            "title": "Diagnostic Framework",
            "type": "string"
          },
          "evaluable_cycle_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Evaluable Cycle Count",
            "type": "integer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_summary": {
            "title": "Evidence Summary",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "meets_chronicity_rule": {
            "default": false,
            "title": "Meets Chronicity Rule",
            "type": "boolean"
          },
          "meets_drsp_content_rule": {
            "default": false,
            "title": "Meets Drsp Content Rule",
            "type": "boolean"
          },
          "meets_drsp_impairment_rule": {
            "default": false,
            "title": "Meets Drsp Impairment Rule",
            "type": "boolean"
          },
          "pack_recommended": {
            "default": false,
            "title": "Pack Recommended",
            "type": "boolean"
          },
          "positive_cycle_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Positive Cycle Count",
            "type": "integer"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "rich_pattern_logged": {
            "default": false,
            "title": "Rich Pattern Logged",
            "type": "boolean"
          },
          "scoring_method": {
            "const": "range_of_scale_used_30pct",
            "default": "range_of_scale_used_30pct",
            "title": "Scoring Method",
            "type": "string"
          },
          "status": {
            "enum": [
              "insufficient_data",
              "suppressed",
              "quiet",
              "activation",
              "pattern"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "supporting_cycle_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Supporting Cycle Count",
            "type": "integer"
          },
          "suppressors": {
            "items": {
              "type": "string"
            },
            "title": "Suppressors",
            "type": "array"
          }
        },
        "required": [
          "analyzer_code",
          "analyzer_version",
          "subject_id",
          "generated_at",
          "status",
          "confidence",
          "summary",
          "evidence_summary"
        ],
        "title": "PmddEvaluationResult",
        "type": "object"
      },
...
      "TrackerSettings": {
        "additionalProperties": false,
        "properties": {
          "disabled_tracker_codes": {
            "items": {
              "type": "string"
            },
            "title": "Disabled Tracker Codes",
            "type": "array"
          },
          "enabled_pack_codes": {
            "items": {
              "type": "string"
            },
            "title": "Enabled Pack Codes",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "tracker_preferences": {
            "items": {
              "$ref": "#/components/schemas/TrackerPreference"
            },
            "title": "Tracker Preferences",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "updated_at",
          "schema_version"
        ],
        "title": "TrackerSettings",
        "type": "object"
      },
      "TrackerSettingsResolutionRequest": {
        "additionalProperties": false,
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/TrackerSettings"
          }
        },
        "required": [
          "settings"
        ],
        "title": "TrackerSettingsResolutionRequest",
        "type": "object"
      },
      "TrackerSettingsValidationResult": {
        "additionalProperties": false,
        "properties": {
          "active_catalog": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveTrackerCatalog"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "ok"
        ],
        "title": "TrackerSettingsValidationResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend-only contract skeleton for a local-first Flutter period tracking app. Device clients own primary calculations.",
    "title": "Period",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analyzers/pcos/evaluate": {
      "post": {
        "description": "Evaluate self-tracked observations against the 2023 IEG PCOS feature pattern.\n\nThis endpoint never diagnoses PCOS. It summarises whether self-reported cycle\nirregularity and clinical hyperandrogenism features are present, flags\nsuppressors that confound interpretation, and lists differentials a clinician\nshould rule out.",
        "operationId": "evaluate_pcos_pattern_api_v1_analyzers_pcos_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PcosEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PcosEvaluationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Pcos Pattern",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analyzers/pmdd/evaluate": {
      "post": {
        "description": "Evaluate recent observations for repeated PMDD-style late-luteal patterns.",
        "operationId": "evaluate_pmdd_pattern_api_v1_analyzers_pmdd_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PmddEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PmddEvaluationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Pmdd Pattern",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/contract-compatibility": {
      "post": {
        "description": "Check whether a schema version is accepted by this backend contract.",
        "operationId": "check_compatibility_api_v1_contract_compatibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractCompatibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCompatibilityResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Compatibility",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/tracker-settings/default": {
      "get": {
        "description": "Return default device-owned tracker settings for a subject id.",
        "operationId": "get_default_tracker_settings_api_v1_tracker_settings_default_get",
        "parameters": [
          {
            "in": "query",
            "name": "subject_id",
            "required": false,
            "schema": {
              "default": "subject-demo",
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackerSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Default Tracker Settings",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/tracker-settings/resolve": {
      "post": {
        "description": "Validate tracker settings and resolve active tracker definitions without storing them.",
        "operationId": "resolve_settings_api_v1_tracker_settings_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackerSettingsResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackerSettingsValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Settings",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/validate-local-data-bundle": {
      "post": {
        "description": "Validate a local-first import/export bundle without storing it.",
        "operationId": "validate_bundle_api_v1_validate_local_data_bundle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDataBundleValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDataBundleValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Bundle",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/validate-local-store-snapshot": {
      "post": {
        "description": "Validate Flutter's on-device local store snapshot without storing it.",
        "operationId": "validate_store_snapshot_api_v1_validate_local_store_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalStoreSnapshotValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDataBundleValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Store Snapshot",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/validate-observation": {
      "post": {
        "description": "Validate a single observation against the Atlas tracker registry.",
        "operationId": "validate_observation_api_v1_validate_observation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Observation",
        "tags": [
          "v1"
        ]
      }
    }
  }
}